Add files via upload
Searching is the process of finding the exact location or position of an element . In linear search we find the element by comparing the list of an element with the given data and if I is equal to size then element is not found. so linear search helps to find the exact position of an element.