Add files via upload
Sorting is the process of arranging a list of elements in a particular order (Ascending or Descending ). Insertion sort algorithm arranges a list of elements in a particular order. In insertion sort algorithm, every iteration moves an element from unsorted portion to sorted portion until all the elements are sorted in the list.